# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface SetupStatus {
  googleCredentials: boolean;
  googleSheet: boolean;
  googleDriveFolder: boolean;
  whatsapp: boolean;
  adminPassword: boolean;
  /** True when every required integration is configured */
  allRequired: boolean;
}

export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface TestPdfInput {
  /** @minLength 1 */
  participantId: string;
  /** @minLength 5 */
  filename: string;
  /** Base64 content of a PDF no larger than 10 MB. */
  pdfBase64: string;
}

export interface LabResultRow {
  field: string;
  /** @nullable */
  value: string | null;
  /** @nullable */
  unit: string | null;
  /** @nullable */
  referenceRange: string | null;
}

export interface TestPdfResult {
  messageId: string;
  participantId: string;
  rowCount: number;
  rows: LabResultRow[];
  sheetsSynced: boolean;
}

export type DrivePollerIntervalInputIntervalMs = typeof DrivePollerIntervalInputIntervalMs[keyof typeof DrivePollerIntervalInputIntervalMs];


export const DrivePollerIntervalInputIntervalMs = {
  NUMBER_900000: 900000,
  NUMBER_3600000: 3600000,
  NUMBER_86400000: 86400000,
} as const;

export interface DrivePollerIntervalInput {
  intervalMs: DrivePollerIntervalInputIntervalMs;
}

export interface DrivePollerStatus {
  running: boolean;
  /** @nullable */
  folderId?: string | null;
  intervalMs?: number;
  /** @nullable */
  lastPollTime?: string | null;
  processedCount: number;
  pendingRetryCount: number;
  pendingRetryFiles: string[];
}

export interface DrivePollerIntervalResult {
  drivePoller: DrivePollerStatus;
}

export type ExtractionSuccessStatus = typeof ExtractionSuccessStatus[keyof typeof ExtractionSuccessStatus];


export const ExtractionSuccessStatus = {
  success: 'success',
} as const;

export type ExtractionSuccessSource = typeof ExtractionSuccessSource[keyof typeof ExtractionSuccessSource];


export const ExtractionSuccessSource = {
  whatsapp: 'whatsapp',
  drive: 'drive',
  manual: 'manual',
} as const;

export interface ExtractionSuccess {
  status: ExtractionSuccessStatus;
  participantId: string;
  messageId: string;
  timestamp: string;
  source: ExtractionSuccessSource;
  rows: LabResultRow[];
}

export type ExtractionErrorStatus = typeof ExtractionErrorStatus[keyof typeof ExtractionErrorStatus];


export const ExtractionErrorStatus = {
  error: 'error',
} as const;

export type ExtractionErrorSource = typeof ExtractionErrorSource[keyof typeof ExtractionErrorSource];


export const ExtractionErrorSource = {
  whatsapp: 'whatsapp',
  drive: 'drive',
  manual: 'manual',
} as const;

export interface ExtractionError {
  status: ExtractionErrorStatus;
  participantId: string;
  messageId: string;
  timestamp: string;
  source: ExtractionErrorSource;
  reason: string;
}

export type ExtractionResult = ExtractionSuccess | ExtractionError;

export interface PollerState {
  enabled: boolean;
  /** @nullable */
  lastRunTime: string | null;
  filesFoundLastRun: number;
  filesSkippedLastRun: number;
  pendingRetryCount: number;
  processedCount: number;
}

export interface AdminExtractionsResponse {
  extractions: ExtractionResult[];
  poller: PollerState;
  retrievedAt: string;
}

export type ExtractionEntrySource = typeof ExtractionEntrySource[keyof typeof ExtractionEntrySource];


export const ExtractionEntrySource = {
  whatsapp: 'whatsapp',
  drive: 'drive',
  manual: 'manual',
} as const;

export type ExtractionEntryStatus = typeof ExtractionEntryStatus[keyof typeof ExtractionEntryStatus];


export const ExtractionEntryStatus = {
  success: 'success',
  error: 'error',
} as const;

export interface ExtractionEntry {
  messageId: string;
  participantId: string;
  timestamp: string;
  source: ExtractionEntrySource;
  status: ExtractionEntryStatus;
  /**
     * Number of lab-result rows extracted (null on error)
     * @nullable
     */
  rowCount?: number | null;
  /**
     * Error reason (null on success)
     * @nullable
     */
  reason?: string | null;
}

export interface SchedulerStatus {
  running: boolean;
  timezone: string;
  morningTime: string;
  eveningTime: string;
  activeParticipantCount: number;
}

export interface AdminStatus {
  extractions: ExtractionEntry[];
  drivePoller: DrivePollerStatus;
  scheduler: SchedulerStatus;
  /** ISO timestamp from the server */
  serverTime: string;
}