# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: admin
    description: Trial coordinator admin endpoints
  - name: setup
    description: In-app setup wizard endpoints
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /admin/setup/status:
    get:
      operationId: getSetupStatus
      tags: [setup]
      summary: Get setup wizard status
      description: >
        Returns which integrations are configured. No auth required — needed
        before the admin password is set during initial wizard flow.
      responses:
        "200":
          description: Setup status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupStatus"

  /admin/extractions:
    get:
      operationId: getAdminExtractions
      tags: [admin]
      summary: List all extractions and poller state
      description: >
        Returns all in-memory extraction results and Drive poller state.
        Protected by HTTP Basic Auth (ADMIN_PASSWORD env var).
      security:
        - basicAuth: []
      responses:
        "200":
          description: Extractions and poller state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminExtractionsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Admin endpoint not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/status:
    get:
      operationId: getAdminStatus
      tags: [admin]
      summary: Full system status snapshot
      description: >
        Returns all extraction results (flattened), Drive poller health, and
        scheduler state. Protected by HTTP Basic Auth (ADMIN_PASSWORD env var).
      security:
        - basicAuth: []
      responses:
        "200":
          description: System status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStatus"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Admin endpoint not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/test-upload:
    post:
      operationId: processTestPdf
      tags: [admin]
      summary: Process a coordinator test PDF
      description: >
        Processes a PDF through the normal extraction and Google Sheets sync path.
        Protected by HTTP Basic Auth.
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestPdfInput"
      responses:
        "200":
          description: PDF processed and results synced to Google Sheets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestPdfResult"
        "400":
          description: Invalid upload input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: PDF could be interpreted as lab results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/drive-poller/interval:
    put:
      operationId: updateDrivePollerInterval
      tags: [admin]
      summary: Set automatic Drive scan frequency
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DrivePollerIntervalInput"
      responses:
        "200":
          description: Drive watcher restarted with the selected interval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrivePollerIntervalResult"
        "400":
          description: Unsupported interval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    SetupStatus:
      type: object
      required: [googleCredentials, googleSheet, googleDriveFolder, whatsapp, adminPassword, allRequired]
      properties:
        googleCredentials:
          type: boolean
        googleSheet:
          type: boolean
        googleDriveFolder:
          type: boolean
        whatsapp:
          type: boolean
        adminPassword:
          type: boolean
        allRequired:
          type: boolean
          description: True when every required integration is configured

    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    TestPdfInput:
      type: object
      required: [participantId, filename, pdfBase64]
      properties:
        participantId:
          type: string
          minLength: 1
        filename:
          type: string
          minLength: 5
        pdfBase64:
          type: string
          description: Base64 content of a PDF no larger than 10 MB.

    TestPdfResult:
      type: object
      required: [messageId, participantId, rowCount, rows, sheetsSynced]
      properties:
        messageId:
          type: string
        participantId:
          type: string
        rowCount:
          type: number
        rows:
          type: array
          items:
            $ref: "#/components/schemas/LabResultRow"
        sheetsSynced:
          type: boolean

    DrivePollerIntervalInput:
      type: object
      required: [intervalMs]
      properties:
        intervalMs:
          type: number
          enum: [900000, 3600000, 86400000]

    DrivePollerIntervalResult:
      type: object
      required: [drivePoller]
      properties:
        drivePoller:
          $ref: "#/components/schemas/DrivePollerStatus"

    # ── /admin/extractions schemas ────────────────────────────────────────────

    LabResultRow:
      type: object
      properties:
        field:
          type: string
        value:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        referenceRange:
          type: ["string", "null"]
      required:
        - field
        - value
        - unit
        - referenceRange

    ExtractionSuccess:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        participantId:
          type: string
        messageId:
          type: string
        timestamp:
          type: string
        source:
          type: string
          enum: [whatsapp, drive, manual]
        rows:
          type: array
          items:
            $ref: "#/components/schemas/LabResultRow"
      required:
        - status
        - participantId
        - messageId
        - timestamp
        - source
        - rows

    ExtractionError:
      type: object
      properties:
        status:
          type: string
          enum: [error]
        participantId:
          type: string
        messageId:
          type: string
        timestamp:
          type: string
        source:
          type: string
          enum: [whatsapp, drive, manual]
        reason:
          type: string
      required:
        - status
        - participantId
        - messageId
        - timestamp
        - source
        - reason

    ExtractionResult:
      oneOf:
        - $ref: "#/components/schemas/ExtractionSuccess"
        - $ref: "#/components/schemas/ExtractionError"

    PollerState:
      type: object
      properties:
        enabled:
          type: boolean
        lastRunTime:
          type: ["string", "null"]
        filesFoundLastRun:
          type: number
        filesSkippedLastRun:
          type: number
        pendingRetryCount:
          type: number
        processedCount:
          type: number
      required:
        - enabled
        - lastRunTime
        - filesFoundLastRun
        - filesSkippedLastRun
        - pendingRetryCount
        - processedCount

    AdminExtractionsResponse:
      type: object
      properties:
        extractions:
          type: array
          items:
            $ref: "#/components/schemas/ExtractionResult"
        poller:
          $ref: "#/components/schemas/PollerState"
        retrievedAt:
          type: string
      required:
        - extractions
        - poller
        - retrievedAt

    # ── /admin/status schemas ─────────────────────────────────────────────────

    ExtractionEntry:
      type: object
      required: [messageId, participantId, timestamp, source, status]
      properties:
        messageId:
          type: string
        participantId:
          type: string
        timestamp:
          type: string
        source:
          type: string
          enum: [whatsapp, drive, manual]
        status:
          type: string
          enum: [success, error]
        rowCount:
          type: ["number", "null"]
          description: Number of lab-result rows extracted (null on error)
        reason:
          type: ["string", "null"]
          description: Error reason (null on success)

    DrivePollerStatus:
      type: object
      required: [running, processedCount, pendingRetryCount, pendingRetryFiles]
      properties:
        running:
          type: boolean
        folderId:
          type: ["string", "null"]
        intervalMs:
          type: number
        lastPollTime:
          type: ["string", "null"]
        processedCount:
          type: number
        pendingRetryCount:
          type: number
        pendingRetryFiles:
          type: array
          items:
            type: string

    SchedulerStatus:
      type: object
      required: [running, timezone, morningTime, eveningTime, activeParticipantCount]
      properties:
        running:
          type: boolean
        timezone:
          type: string
        morningTime:
          type: string
        eveningTime:
          type: string
        activeParticipantCount:
          type: number

    AdminStatus:
      type: object
      required: [extractions, drivePoller, scheduler, serverTime]
      properties:
        extractions:
          type: array
          items:
            $ref: "#/components/schemas/ExtractionEntry"
        drivePoller:
          $ref: "#/components/schemas/DrivePollerStatus"
        scheduler:
          $ref: "#/components/schemas/SchedulerStatus"
        serverTime:
          type: string
          description: ISO timestamp from the server