# lib/api-zod/src/generated/types/drivePollerIntervalResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DrivePollerStatus } from './drivePollerStatus';

export interface DrivePollerIntervalResult {
  drivePoller: DrivePollerStatus;
}