# lib/api-zod/src/generated/types/testPdfInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TestPdfInput {
  /** @minLength 1 */
  participantId: string;
  /** @minLength 5 */
  filename: string;
  /** Base64 content of a PDF no larger than 10 MB. */
  pdfBase64: string;
}