# lib/api-zod/src/generated/types/testPdfResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LabResultRow } from './labResultRow';

export interface TestPdfResult {
  messageId: string;
  participantId: string;
  rowCount: number;
  rows: LabResultRow[];
  sheetsSynced: boolean;
}